Add files via upload
This contains a code for the WhatsApp blaster to be integrated with ERP softwares for real-time updates to clients and customers on WhatsApp. This will help the ERP company automate the process of attachment of documents and images on the message panel to cell phone numbers without actually saving them in the phone directory.